+++ local-server/src/certificates/mod.rs
@@ -1,120 +1,36 @@
+mod wildcard_sni_resolver;
+
 use rcgen::{Certificate, CertificateParams, DistinguishedName, DnType, KeyPair};
 use rustls::crypto::ring::sign;
 use rustls::pki_types::CertificateDer;
-use rustls::server::{ClientHello, ResolvesServerCert};
 use rustls::sign::CertifiedKey;
-use std::collections::HashMap;
-use std::sync::{Arc, RwLock};
-use std::{env, fs, path::PathBuf, process};
+use std::{
+    env, fs,
+    path::{Path, PathBuf},
+    process,
+};
+
+pub use wildcard_sni_resolver::WildcardSniResolver;
 
 const LINKUP_CA_COMMON_NAME: &str = "Linkup Local CA";
 
-pub fn ca_cert_pem_path(certs_dir: &PathBuf) -> PathBuf {
+pub fn ca_cert_pem_path(certs_dir: &Path) -> PathBuf {
     certs_dir.join("linkup_ca.cert.pem")
 }
 
-pub fn ca_key_pem_path(certs_dir: &PathBuf) -> PathBuf {
+pub fn ca_key_pem_path(certs_dir: &Path) -> PathBuf {
     certs_dir.join("linkup_ca.key.pem")
 }
 
-#[derive(Debug)]
-pub struct WildcardSniResolver {
-    certs: RwLock<HashMap<String, Arc<CertifiedKey>>>,
-}
-
-impl WildcardSniResolver {
-    fn new() -> Self {
-        Self {
-            certs: RwLock::new(HashMap::new()),
-        }
-    }
-
-    fn add_cert(&self, domain: &str, cert: CertifiedKey) {
-        let mut certs = self.certs.write().unwrap();
-        certs.insert(domain.to_string(), Arc::new(cert));
-    }
-
-    fn find_cert(&self, server_name: &str) -> Option<Arc<CertifiedKey>> {
-        let certs = self.certs.read().unwrap();
-
-        if let Some(cert) = certs.get(server_name) {
-            return Some(cert.clone());
-        }
-
-        let parts: Vec<&str> = server_name.split('.').collect();
-
-        for i in 0..parts.len() {
-            let wildcard_domain = format!("*.{}", parts[i..].join("."));
-            if let Some(cert) = certs.get(&wildcard_domain) {
-                return Some(cert.clone());
-            }
-        }
-
-        None
-    }
-}
-
-impl ResolvesServerCert for WildcardSniResolver {
-    fn resolve(&self, client_hello: ClientHello<'_>) -> Option<Arc<CertifiedKey>> {
-        if let Some(name) = client_hello.server_name() {
-            return self.find_cert(name.as_ref());
-        }
-
-        None
-    }
-}
-
-pub fn load_certificates_from_dir(cert_dir: &PathBuf) -> WildcardSniResolver {
-    let resolver = WildcardSniResolver::new();
-
-    let entries = fs::read_dir(cert_dir).expect("Failed to read certs directory");
-
-    for entry in entries.flatten() {
-        let path = entry.path();
-
-        if path
-            .file_name()
-            .unwrap()
-            .to_string_lossy()
-            .contains(".cert.pem")
-            && !path.starts_with("linkup_ca")
-        {
-            let path_str = path.to_string_lossy();
-            let domain_name = path
-                .file_name()
-                .unwrap()
-                .to_string_lossy()
-                .replace(".cert.pem", "")
-                .replace("wildcard_", "*");
-            let key_path = PathBuf::from(path_str.replace(".cert.pem", ".key.pem"));
-
-            if key_path.exists() {
-                match load_cert_and_key(path, key_path) {
-                    Ok(certified_key) => {
-                        println!("Loaded certificate for {}", domain_name);
-                        resolver.add_cert(&domain_name, certified_key);
-                    }
-                    Err(e) => {
-                        eprintln!("Error loading cert/key for {domain_name}: {e}");
-                    }
-                }
-            }
-        }
-    }
-
-    resolver
-}
-
 fn load_cert_and_key(
-    cert_path: PathBuf,
-    key_path: PathBuf,
+    cert_path: &Path,
+    key_path: &Path,
 ) -> Result<CertifiedKey, Box<dyn std::error::Error>> {
     let cert_pem = fs::read(cert_path)?;
     let key_pem = fs::read(key_path)?;
 
     let certs = rustls_pemfile::certs(&mut &cert_pem[..])
         .filter_map(|cert| cert.ok())
-        .map(CertificateDer::from)
         .collect::<Vec<CertificateDer<'static>>>();
 
     if certs.is_empty() {
@@ -134,7 +50,7 @@ fn load_cert_and_key(
     })
 }
 
-pub fn create_domain_cert(certs_dir: &PathBuf, domain: &str) -> (Certificate, KeyPair) {
+pub fn create_domain_cert(certs_dir: &Path, domain: &str) -> (Certificate, KeyPair) {
     let cert_pem_str = fs::read_to_string(ca_cert_pem_path(certs_dir)).unwrap();
     let key_pem_str = fs::read_to_string(ca_key_pem_path(certs_dir)).unwrap();
 
@@ -156,12 +72,15 @@ pub fn create_domain_cert(certs_dir: &PathBuf, domain: &str) -> (Certificate, Ke
     fs::write(cert_path, cert.pem()).unwrap();
     fs::write(key_path, key_pair.serialize_pem()).unwrap();
 
-    println!("Certificate for {} generated!", domain);
-
     (cert, key_pair)
 }
 
-pub fn upsert_ca_cert(certs_dir: &PathBuf) {
+pub fn install_ca_certificate(certs_dir: &Path) {
+    upsert_ca_cert(certs_dir);
+    add_ca_to_keychain(certs_dir);
+}
+
+fn upsert_ca_cert(certs_dir: &Path) {
     if ca_cert_pem_path(certs_dir).exists() && ca_key_pem_path(certs_dir).exists() {
         return;
     }
@@ -184,7 +103,7 @@ pub fn upsert_ca_cert(certs_dir: &PathBuf) {
     fs::write(ca_key_pem_path(certs_dir), key_pair.serialize_pem()).unwrap();
 }
 
-pub fn add_ca_to_keychain(certs_dir: &PathBuf) {
+fn add_ca_to_keychain(certs_dir: &Path) {
     process::Command::new("sudo")
         .arg("security")
         .arg("add-trusted-cert")
@@ -196,7 +115,7 @@ pub fn add_ca_to_keychain(certs_dir: &PathBuf) {
         .arg(ca_cert_pem_path(certs_dir))
         .stdout(process::Stdio::piped())
         .stderr(process::Stdio::piped())
-        .spawn()
+        .status()
         .expect("Failed to add CA to keychain");
 }
 
@@ -206,16 +125,14 @@ pub fn install_nss() {
         return;
     }
 
-    let mut cmd = process::Command::new("brew")
+    process::Command::new("brew")
         .arg("install")
         .arg("nss")
-        .spawn()
+        .status()
         .expect("Failed to install NSS");
-
-    cmd.wait().expect("Failed to wait for NSS install");
 }
 
-pub fn add_ca_to_nss(certs_dir: &PathBuf) {
+pub fn add_ca_to_nss(certs_dir: &Path) {
     if !is_nss_installed() {
         println!("NSS not found, skipping CA installation");
         return;
@@ -230,9 +147,9 @@ pub fn add_ca_to_nss(certs_dir: &PathBuf) {
                 let path = entry.path();
                 if path.is_dir() {
                     if path.join("cert9.db").exists() {
-                        return Some(format!("{}{}", "sql:", path.to_str().unwrap()));
+                        Some(format!("{}{}", "sql:", path.to_str().unwrap()))
                     } else if path.join("cert8.db").exists() {
-                        return Some(format!("{}{}", "dmb:", path.to_str().unwrap()));
+                        Some(format!("{}{}", "dmb:", path.to_str().unwrap()))
                     } else {
                         None
                     }
@@ -253,7 +170,7 @@ pub fn add_ca_to_nss(certs_dir: &PathBuf) {
             .arg(LINKUP_CA_COMMON_NAME)
             .arg("-i")
             .arg(ca_cert_pem_path(certs_dir))
-            .spawn()
+            .status()
             .expect("Failed to add CA to NSS");
     }
 }
